==> code/01_cleaning.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from ds_tools import cleaner

pd.options.display.max_rows = 5000
pd.options.display.max_columns = 5000

train_df = pd.read_csv('../datasets/train.csv')


pd.Da

clean_df = cleaner(train_df)

# change whole df to lower and proper spacing

train_df.columns = train_df.columns.str.lower().str.replace(' ', '_') 

train_df = train_df.applymap(lambda x : x.lower() if isinstance(x, str) else x)

train_df.sample(5)

# saving font and spacing correction
train_df.to_csv('../datasets/train.csv', index=False)

def find_NA(df):
    na_mean = df.isna().mean()[df.isna().mean().sort_values() > 0].sort_values()
    na_total = df.isna().sum()[df.isna().sum().sort_values() > 0].sort_values() 
    
    na_df = pd.DataFrame(na_total, columns=['total_nulls'])
    na_df['mean_nulls'] = na_mean
    na_df['total_count'] = df.isna().count()
    na_df['dtypes'] = train_df.dtypes
    
    return na_df
    

null_df = find_NA(train_df)

null_df

train_df = train_df.drop(['pool_qc', 'misc_feature', 'fence', 'alley', 'lot_frontage', 'id', 'pid'], axis=1) 

# null for fireplace == no fireplace

train_df['fireplace_qu'] = train_df['fireplace_qu'].fillna('none')

fill_na = ['garage_type', 'garage_finish', 'garage_qual', 'garage_cond', 'Fireplace Qu', 
           'Bsmt Qual', 'Bsmt Cond', 'Bsmt Exposure', 'BsmtFin Type 1', 
           'BsmtFin Type 2', 'Mas Vnr Type', 'Bsmt Qual', 'Bsmt Cond', 'Bsmt Exposure', 'BsmtFin Type 1',
           'BsmtFin SF 1', 'BsmtFin Type 2', 'BsmtFin SF 2', 'Bsmt Unf SF',
           'Bsmt Full Bath', 'Bsmt Half Bath', 'mas_vnr_area', 'total_bsmt_sf', 'garage_cars', 'garage_area']

fill_na = [feature.lower().replace(' ', '_') for feature in fill_na] 

train_df[fill_na] = train_df[fill_na].fillna(0)

# The year built and garage_yr_blt are closely related
g_yr_na = (train_df['year_built'] - train_df['garage_yr_blt']).mean()

# Most are built the same year, with an avg of with -5 yrear

# filling all null value with year_built - mean 
train_df['garage_yr_blt'] = train_df['garage_yr_blt'].fillna(train_df['year_built'] - g_yr_na) 

find_NA(train_df)

train_df.to_csv('../datasets/train_clean.csv', index=False)
