==> code/04_Modeling.ipynb <==
import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler, PolynomialFeatures
import sklearn.metrics

from sklearn.model_selection import train_test_split, KFold, cross_val_score
from sklearn.linear_model import Ridge, Lasso, LinearRegression, ElasticNet

from sklearn.pipeline import Pipeline

class lr_model:
    'Modeling Class, default to LinearRegression'
    def __init__(self, df, target, pipe_steps):
        self.df = df
        self.target = self.df[target]
        self.features = self.df.drop(target, axis=1)
        
        self.num_features = self.features.select_dtypes(include='number')
        self.nom_features = self.features.select_dtypes(exclude='number')
        
        self.pipe = Pipeline(pipe_steps)
        
        self.summary = pd.DataFrame({'random_state' : [], 'val_score': [], 'train_score' : [], 'test_score' : []})
        
    def test_models(self, run_time=3):
        'Run Models X amount of time with different random state'
        for i in np.random.choice(100, run_time, replace=False):
            seed = i
            X = self.features
            y = self.target

            X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=seed) 
            self.pipe.fit(X_train, y_train)

            val_score = round(cross_val_score(self.pipe, X_train, y_train, cv=5).mean(), 2)
            test_score = round(self.pipe.score(X_test, y_test), 2)
            train_score = round(self.pipe.score(X_train, y_train), 2)

            self.summary = self.summary.append({'random_state' : i, 'val_score': val_score, 
                                 'train_score' : train_score, 'test_score' : test_score}, ignore_index=True)
        
        return self.summary
    
    def final_model(self, seed):
        X = self.features
        y = self.target

        X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=seed) 
        self.pipe.fit(X_train, y_train)
    
    def predictions(self, data):
        return self.pipe.predict(data)
                
        
    def coef_score(self):
        
        X = self.features
        y = self.target
        X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=8) 
        self.model.fit(X_train, y_train)
        
                    
        val_score = cross_val_score(self.model, X_train, y_train, cv=5).mean()
        test_score = self.model.score(X_test, y_test)
        train_score = self.model.score(X_train, y_train)
        
        
        summary = pd.DataFrame({"coefficients": np.transpose([(round(coef, 2)) for coef in self.model.coef_]), 
                             'avg_feature_value' : self.features.mean(), 
                             'avg_feature__median' : self.features.median()})
        
        summary['avg_change'] = summary['avg_feature_value'] * summary['coefficients']
        summary['count'] = self.features[self.features > 0].count()
        
        return summary.sort_values('avg_change', ascending=False)
    

df = pd.read_csv('../datasets/model_data.csv')

lr_mysteps = [
    ('sc', StandardScaler()),
    ('lr', LinearRegression()) # estimator         
    ]

lr_model = lr_model(df, 'saleprice', lr_mysteps)

lr_model.test_models(10)

ridge_mysteps = [
    ('sc', StandardScaler()),
    ('lr', Ridge()) # estimator         
    ]

ridge_model = lr_model(df, 'saleprice', ridge_mysteps)

ridge_model.test_models(10)

lasso_mysteps = [
    ('sc', StandardScaler()),
    ('lr', Lasso()) # estimator         
    ]

lasso_model = lr_model(df, 'saleprice', lasso_mysteps)

lasso_model.test_models(10)

el_mysteps = [
    ('sc', StandardScaler()),
    ('lr', ElasticNet()) # estimator         
    ]

el_model = lr_model(df, 'saleprice', el_mysteps)

el_model.test_models(10)
